requirements.txt
requests
Unidecode
pandas
python-slugify